Remove deprecated settings implementation
This removes the old settings module and renames setting2.py to
setting.py. This is a breaking change for plugins that were doing
something like `from armi.settings import setting2 as setting`, but
trivial to refactor around.
