+++ armi/settings/fwSettings/databaseSettings.py
@@ -14,7 +14,7 @@
 
 """Settings related to the ARMI database."""
 
-from armi.settings import setting2 as setting
+from armi.settings import setting
 
 
 CONF_DB = "db"
